재고수량부족 예외처리, 테스트코드 추가

+++ home-inventory-management/src/main/java/me/mason/management/application/UseInventory.java
@@ -32,7 +32,7 @@ public class UseInventory implements UseInventoryInput {
         Inventory inventory = modifyInventoryStateOutput.loadByInventoryId(inventoryId);
 
         //재고 사용처리
-        inventory.use(quantity.intValue());
+        inventory.use(quantity);
 
         modifyInventoryStateOutput.modifyInventory(inventory);
